finish responsee type converter with typing
