+++ src/cqr-event/response-type.ts
@@ -17,10 +17,6 @@
  *   limitations under the License.
  *
  */
-import type { MetaActionCreator } from '../classify/meta-action-creator.js'
-
-import type { MetaResponse }  from './meta.js'
-
 /**
  * We add `_RESPONSE` to the end of the `type` for the Response Event for Command/Query
  */
@@ -28,6 +24,9 @@ const _RESPONSE = '_RESPONSE'
 export type ResponseType<T extends string> = `${T}${typeof _RESPONSE}`
 export const responseType = <T extends string> (type: T) => `${type}${_RESPONSE}` as ResponseType<T>
 
+/**
+ * Deal with a map object {}
+ */
 export type ResponseTypeMap<T extends { [key: string]: string }> = {
   [K in keyof T as ResponseType<string & K>]: ResponseType<T[K]>
 }
@@ -38,27 +37,3 @@ export const responseTypeMap = <T extends { [key: string]: string }> (o: T) =>
     acc[k] = v
     return acc
   }, {} as any) as ResponseTypeMap<T>
-
-/**
- * Huan(202203): FIXME: use the real `Symbol` instead of a `string`
- *
- * Error message:
- *
- *  Exported variable 'resetCommand' has or is using name 'RESPONSE'
- *  from external module "/home/huan/git/wechaty/cqrs/src/duck/actions/create-action-pair"
- *  but cannot be named. ts(4023)
- */
-
-export const RESPONSE = "Symbol('RESPONSE')"
-export interface Responseable <
-  R extends MetaActionCreator<string, {}, MetaResponse> = MetaActionCreator<string, {}, MetaResponse>,
-> {
-  [RESPONSE]: R
-}
-
-export type ResponseOf<T extends Responseable> = T[typeof RESPONSE]
-// Huan(202203) the tailing `as T[...]` is not necessary in theory but it seems to be a bug in TS
-export const responseOf = <T extends Responseable> (responseable: T) => responseable[RESPONSE] as T[typeof RESPONSE]
-
-export type ActionOf<T extends Responseable> = Omit<T, typeof RESPONSE>
-export const actionOf = <T extends Responseable>(action: T): ActionOf<T> => action
